table-segmentation v5
Based on v4 . The code is completely refactored with correct software patterns.

The script can now receive any of this inputs : an image or a pdf, a folder with a set of images or pdfs. It will treat all the given documents and output it to the specified output directory. 

What the script does : The script extracts/crops a table of a page or image using a basic line detection or contour-based detection, so i can be better treated afterwards.
What the script does not : The script does not make "cell recognition", "character recognition" or any data extraction.

```
python table_extractor.py <chemin_vers_le_fichier> <chemin_de_sortie>
```
Changes :
 - Extracted user defined variables to a config.py file
 - Added a CLI for easy use in command line :  table_extractor.py
 - Added a README.md with basic usage instructions, dependencies used and credits.
 - Added a requirements.txt for easier installation.
 - Main functionalities are contained inside extractor_module.py (No changes on behaviour).